backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.8.1
setuptools==75.8.0
python-chess==1.999
langgraph==0.3.6
pymilvus==2.4.10
fastembed==0.7.4
